hello-langgraph/pyproject.toml
[project]
name = "hello-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain[openai]>=1.2.13",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.19",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
